+++ Dockerfile
@@ -1,7 +1,9 @@
 # Build the manager binary
 FROM europe-docker.pkg.dev/kyma-project/prod/external/golang:1.21.4-alpine3.18 as builder
-ARG TARGETOS
-ARG TARGETARCH
+ARG TARGET_OS
+ARG TARGET_ARCH
+ARG VERSION
+ARG BUILD_TIME
 
 WORKDIR /workspace
 # Copy the Go Modules manifests
@@ -13,7 +15,6 @@ RUN go mod download
 
 # Copy the go source
 COPY main.go main.go
-COPY VERSION VERSION
 COPY apis/ apis/
 COPY controllers/ controllers/
 COPY internal/ internal/
@@ -24,7 +25,7 @@ COPY manifests/ manifests/
 # was called. For example, if we call make docker-build in a local env which has the Apple Silicon M1 SO
 # the docker BUILDPLATFORM arg will be linux/arm64 when for Apple x86 it will be linux/amd64. Therefore,
 # by leaving it empty we can ensure that the container and binary shipped on it will have the same platform.
-RUN CGO_ENABLED=0 GOOS=${TARGETOS:-linux} GOARCH=${TARGETARCH:-amd64} go build -a -o manager main.go
+RUN CGO_ENABLED=0 GOOS=${TARGET_OS:-linux} GOARCH=${TARGET_ARCH:-amd64} go build -ldflags="-X 'github.com/kyma-project/api-gateway/internal/version.version=${VERSION:-dev}' -X 'github.com/kyma-project/api-gateway/internal/version.buildTime=${BUILD_TIME:-}'" -a -o manager main.go
 
 
 # Use distroless as minimal base image to package the manager binary
@@ -32,7 +33,6 @@ RUN CGO_ENABLED=0 GOOS=${TARGETOS:-linux} GOARCH=${TARGETARCH:-amd64} go build -
 FROM gcr.io/distroless/static:nonroot
 WORKDIR /
 COPY --from=builder /workspace/manager .
-COPY --from=builder /workspace/VERSION .
 COPY --from=builder /workspace/manifests/ manifests
 
 USER 65532:65532
