--- VERSION
@@ -1,1 +1,0 @@
-2.1.0
