+++ docs/contributor/04-10-technical-design.md
@@ -38,28 +38,3 @@ In the event of a failure during the reconciliation, APIRule Controller performs
 The following diagram illustrates the reconciliation process of APIRule and the created resources:
 
 ![APIRule CR Reconciliation](../assets/api-rule-reconciliation-sequence.svg)
-
-## Labeling resources
-
-Based on the decision taken for [Consistent labeling of Kyma modules](https://github.com/kyma-project/community/issues/864). APIGateway operator resources are labeled with the common kubernetes labels:
-
-```yaml
-kyma-project.io/module: api-gateway
-app.kubernetes.io/name: api-gateway-operator
-app.kubernetes.io/instance: api-gateway-operator-default
-app.kubernetes.io/version: "x.x.x"
-app.kubernetes.io/component: operator
-app.kubernetes.io/part-of: api-gateway
-```
-
-For all other resource like the external `ory-oathkeeper` component and its respective resources we only apply the Kyma module label:
-
-```yaml
-kyma-project.io/module: api-gateway
-```
-
-Getting all resources created by the API Gateway module is possible by running:
-
-```bash
-kubectl get all|<resources-kind> -A -l kyma-project.io/module=api-gateway
-```
