+++ docs/user/00-10-overview-api-gateway-controller.md
@@ -21,3 +21,28 @@ The `apigateways.operator.kyma-project.io` CustomResourceDefinition (CRD) descri
 |  `Deleting`  | Controller is deleting resources.        |
 |   `Error`    | An error occurred during reconciliation. |
 |  `Warning`   | Controller is misconfigured.             |
+
+## Labeling resources
+
+Based on the decision taken for [Consistent labeling of Kyma modules](https://github.com/kyma-project/community/issues/864). APIGateway operator resources are labeled with the common kubernetes labels:
+
+```yaml
+kyma-project.io/module: api-gateway
+app.kubernetes.io/name: api-gateway-operator
+app.kubernetes.io/instance: api-gateway-operator-default
+app.kubernetes.io/version: "x.x.x"
+app.kubernetes.io/component: operator
+app.kubernetes.io/part-of: api-gateway
+```
+
+For all other resource like the external `ory-oathkeeper` component and its respective resources we only apply the Kyma module label:
+
+```yaml
+kyma-project.io/module: api-gateway
+```
+
+Getting all resources created by the API Gateway module is possible by running:
+
+```bash
+kubectl get all|<resources-kind> -A -l kyma-project.io/module=api-gateway
+```
