+++ internal/helpers/service_test.go
@@ -2,12 +2,9 @@ package helpers
 
 import (
 	"context"
-	"errors"
-	"fmt"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"golang.org/x/mod/semver"
 
 	gatewayv1beta1 "github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
 	corev1 "k8s.io/api/core/v1"
@@ -173,23 +170,5 @@ var _ = Describe("GetLabelSelectorFromService", func() {
 		workloadSelector, err := GetLabelSelectorFromService(context.TODO(), k8sClient, &service, &apiRule, &rule)
 		Expect(err).ToNot(HaveOccurred())
 		Expect(workloadSelector).To(BeNil())
-	})
-})
-
-var _ = Describe("GetModuleVersion", func() {
-	It("Should return a valid version if succesfully read", func() {
-		ReadVersionFileHandle = func(name string) ([]byte, error) {
-			return []byte("1.0.0"), nil
-		}
-		version := GetModuleVersion()
-		Expect(semver.IsValid(fmt.Sprintf("v%s", version))).To(BeTrue())
-	})
-
-	It("Should return unknown version if could not read it", func() {
-		ReadVersionFileHandle = func(name string) ([]byte, error) {
-			return []byte{}, errors.New("could not read")
-		}
-		version := GetModuleVersion()
-		Expect(version).To(Equal("unknown"))
 	})
 })
