+++ internal/reconciliations/gateway/kyma_gateway.go
@@ -6,139 +6,64 @@ import (
 	"fmt"
 
 	"github.com/kyma-project/api-gateway/apis/operator/v1alpha1"
-	"github.com/kyma-project/api-gateway/controllers"
 	"github.com/kyma-project/api-gateway/internal/reconciliations"
-	"github.com/kyma-project/api-gateway/internal/resources"
+	"github.com/kyma-project/api-gateway/internal/version"
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	ctrl "sigs.k8s.io/controller-runtime"
+
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
-	nonGardenerDomainName = "local.kyma.dev"
+	KymaGatewayName      = "kyma-gateway"
+	KymaGatewayNamespace = "kyma-system"
+	KymaGatewayFullName  = "kyma-system/kyma-gateway"
 )
 
-var checkDefaultGatewayReference = func(ctx context.Context, c client.Client, res resources.Resource) bool {
-	u := &unstructured.Unstructured{}
-	u.SetGroupVersionKind(res.GVK)
+//go:embed kyma_gateway.yaml
+var kymaGatewayManifest []byte
 
-	err := c.Get(ctx, client.ObjectKey{
-		Namespace: res.Namespace,
-		Name:      res.Name,
-	}, u)
+func reconcileKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway, domain string) error {
+	isEnabled := isKymaGatewayEnabled(apiGatewayCR)
+	ctrl.Log.Info("Reconciling Kyma gateway", "KymaGatewayEnabled", isEnabled)
 
+	templateValues := make(map[string]string)
+	templateValues["Name"] = KymaGatewayName
+	templateValues["Namespace"] = KymaGatewayNamespace
+	templateValues["Domain"] = domain
+	templateValues["CertificateSecretName"] = kymaGatewayCertSecretName
+	templateValues["Version"] = version.GetVersion()
+
+	resource, err := reconciliations.CreateUnstructuredResource(kymaGatewayManifest, templateValues)
 	if err != nil {
-		ctrl.Log.Error(err, "Error happened during getting object")
+		return err
 	}
 
-	if res.GVK.Kind == "APIRule" && u.Object["spec"] != nil {
-		return u.Object["spec"].(map[string]interface{})["gateway"] == KymaGatewayFullName
-	} else if res.GVK.Kind == "VirtualService" && u.Object["spec"] != nil {
-		gateways := u.Object["spec"].(map[string]interface{})["gateways"]
-		if gateways != nil {
-			for _, gateway := range gateways.([]interface{}) {
-				if gateway == KymaGatewayFullName {
-					return true
-				}
-			}
-		}
+	if !isEnabled || apiGatewayCR.IsInDeletion() {
+		return deleteKymaGateway(ctx, k8sClient, resource)
 	}
 
-	return false
+	return reconciliations.CreateOrUpdateResource(ctx, k8sClient, resource)
 }
 
-// ReconcileKymaGateway reconciles the kyma-gateway and creates all required resources for the Gateway to fully work. It also adds a finalizer to
-// APIGateway CR and handles the deletion of the resources if the APIGateway CR is deleted.
-// Returns a Status object with the result of the reconciliation and an error if the reconciliation failed.
-func ReconcileKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR *v1alpha1.APIGateway, apiGatewayResourceListPath string) controllers.Status {
-	ctrl.Log.Info("Reconcile Kyma Gateway", "enabled", apiGatewayCR.Spec.EnableKymaGateway)
-	if isKymaGatewayEnabled(*apiGatewayCR) && !apiGatewayCR.IsInDeletion() && !hasKymaGatewayFinalizer(*apiGatewayCR) {
-		if err := addKymaGatewayFinalizer(ctx, k8sClient, apiGatewayCR); err != nil {
-			return controllers.ErrorStatus(err, "Failed to add finalizer during Kyma Gateway reconciliation")
-		}
+func isKymaGatewayEnabled(cr v1alpha1.APIGateway) bool {
+	return cr.Spec.EnableKymaGateway != nil && *cr.Spec.EnableKymaGateway
+}
+
+func deleteKymaGateway(ctx context.Context, k8sClient client.Client, kymaGateway unstructured.Unstructured) error {
+	ctrl.Log.Info("Deleting Kyma gateway")
+	err := k8sClient.Delete(ctx, &kymaGateway)
+
+	if err != nil && !errors.IsNotFound(err) {
+		return fmt.Errorf("failed to delete Kyma gateway: %v", err)
 	}
 
-	if !hasKymaGatewayFinalizer(*apiGatewayCR) {
-		ctrl.Log.Info("There is no Kyma Gateway finalizer, skipping reconciliation")
-		return controllers.ReadyStatus()
-	}
-
-	if !isKymaGatewayEnabled(*apiGatewayCR) || apiGatewayCR.IsInDeletion() {
-		resourceFinder, err := resources.NewResourcesFinderFromConfigYaml(ctx, k8sClient, ctrl.Log, apiGatewayResourceListPath)
-		if err != nil {
-			return controllers.ErrorStatus(err, "Could not read customer resources finder configuration")
-		}
-
-		clientResources, err := resourceFinder.FindUserCreatedResources(checkDefaultGatewayReference)
-		if err != nil {
-			return controllers.ErrorStatus(err, "Could not get customer resources from the cluster")
-		}
-
-		if len(clientResources) > 0 {
-			for _, res := range clientResources {
-				ctrl.Log.Info("Custom resource is blocking Kyma Gateway deletion", "gvk", res.GVK.String(), "namespace", res.Namespace, "name", res.Name)
-			}
-
-			return controllers.WarningStatus(fmt.Errorf("could not delete Kyma Gateway since there are %d custom resource(s) present that block its deletion", len(clientResources)),
-				"There are custom resources that block the deletion of Kyma Gateway. Please take a look at kyma-system/api-gateway-controller-manager logs to see more information about the warning")
-		}
-	}
-
-	isGardenerCluster, err := reconciliations.RunsOnGardenerCluster(ctx, k8sClient)
-	if err != nil {
-		return controllers.ErrorStatus(err, "Error during Kyma Gateway reconciliation")
-	}
-
-	var reconcileErr error
-	if isGardenerCluster {
-		reconcileErr = reconcileGardenerKymaGateway(ctx, k8sClient, *apiGatewayCR)
+	if errors.IsNotFound(err) {
+		ctrl.Log.Info("Skipped deletion of Kyma gateway as it wasn't present")
 	} else {
-		reconcileErr = reconcileNonGardenerKymaGateway(ctx, k8sClient, *apiGatewayCR)
+		ctrl.Log.Info("Successfully deleted Kyma gateway")
 	}
 
-	if reconcileErr != nil {
-		return controllers.ErrorStatus(reconcileErr, "Error during Kyma Gateway reconciliation")
-	}
-
-	// Besides on disabling the Kyma gateway, we also need to remove the finalizer on APIGateway deletion to make sure we are not blocking the deletion of the CR.
-	if !isKymaGatewayEnabled(*apiGatewayCR) || apiGatewayCR.IsInDeletion() {
-		if err := removeKymaGatewayFinalizer(ctx, k8sClient, apiGatewayCR); err != nil {
-			return controllers.ErrorStatus(err, "Failed to remove finalizer during Kyma Gateway reconciliation")
-		}
-	}
-
-	return controllers.ReadyStatus()
-}
-
-func reconcileGardenerKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway) error {
-	domain, err := reconciliations.GetGardenerDomain(ctx, k8sClient)
-	if err != nil {
-		return fmt.Errorf("failed to get Kyma gateway domain: %v", err)
-	}
-
-	if err := reconcileKymaGatewayDnsEntry(ctx, k8sClient, apiGatewayCR, domain); err != nil {
-		return err
-	}
-
-	if err := reconcileKymaGatewayCertificate(ctx, k8sClient, apiGatewayCR, domain); err != nil {
-		return err
-	}
-
-	if err := reconcileKymaGatewayVirtualService(ctx, k8sClient, apiGatewayCR, domain); err != nil {
-		return err
-	}
-
-	return reconcileKymaGateway(ctx, k8sClient, apiGatewayCR, domain)
-}
-
-func reconcileNonGardenerKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway) error {
-	if err := reconcileNonGardenerCertificateSecret(ctx, k8sClient, apiGatewayCR); err != nil {
-		return err
-	}
-
-	if err := reconcileKymaGatewayVirtualService(ctx, k8sClient, apiGatewayCR, nonGardenerDomainName); err != nil {
-		return err
-	}
-
-	return reconcileKymaGateway(ctx, k8sClient, apiGatewayCR, nonGardenerDomainName)
+	return nil
 }
