+++ internal/reconciliations/gateway/kyma_gateway.yaml
