+++ internal/reconciliations/gateway/virtual_service.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 
 	"github.com/kyma-project/api-gateway/apis/operator/v1alpha1"
-	"github.com/kyma-project/api-gateway/internal/helpers"
 	"github.com/kyma-project/api-gateway/internal/reconciliations"
+	"github.com/kyma-project/api-gateway/internal/version"
 	"istio.io/client-go/pkg/apis/networking/v1beta1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -41,7 +41,7 @@ func reconcileVirtualService(ctx context.Context, k8sClient client.Client, name,
 	templateValues["Namespace"] = namespace
 	templateValues["Domain"] = domain
 	templateValues["Gateway"] = KymaGatewayFullName
-	templateValues["Version"] = helpers.GetModuleVersion()
+	templateValues["Version"] = version.GetVersion()
 
 	return reconciliations.ApplyResource(ctx, k8sClient, virtualServiceManifest, templateValues)
 }
