+++ main.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	"github.com/kyma-project/api-gateway/internal/reconciliations/oathkeeper"
+	"github.com/kyma-project/api-gateway/internal/version"
 
 	gatewayv1beta1 "github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
 	"github.com/kyma-project/api-gateway/controllers"
@@ -183,6 +184,9 @@ func main() {
 	}
 
 	setupLog.Info("starting manager")
+	setupLog.Info("module version", "version", version.GetVersion())
+	setupLog.Info("build timestamp", "timestamp", version.GetBuildTime())
+
 	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
 		setupLog.Error(err, "problem running manager")
 		os.Exit(1)
